Metropolis: Rueckgabewert der Methoden einkommensteuer(),koerperschaftsteuer() usw auf int abgeaendert.Methode einkommensteuer() verbessert.

+++ 04_objektorientierung/13_solid/src/tpe/oo/metropolis/Einkommensteuerpflichtig.java
@@ -5,5 +5,5 @@ public interface Einkommensteuerpflichtig extends Steuerpflichtig {
     double STEUERSATZ2=0.15;
     double STEUERSATZ3=0.25;
 
-    double einkommensteuer();
+    int einkommensteuer();
 }
