+++ 04_objektorientierung/13_solid/src/tpe/oo/metropolis/Einwohner.java
@@ -2,9 +2,9 @@ package tpe.oo.metropolis;
 
 public abstract class Einwohner {
 protected String name;
-protected double einkommen;
+protected int einkommen;
 
-protected Einwohner(String name,double einkommen){
+protected Einwohner(String name,int einkommen){
     this.name = name;
     this.einkommen= einkommen;
 }
