+++ 04_objektorientierung/13_solid/src/tpe/oo/metropolis/Koerperschaftsteuerpflichtig.java
@@ -2,5 +2,5 @@ package tpe.oo.metropolis;
 
 public interface Koerperschaftsteuerpflichtig extends Steuerpflichtig {
     double STEUERSATZ = 0.15;
-    double koerperschaftsteuer();
+    int koerperschaftsteuer();
 }
