+++ 04_objektorientierung/13_solid/src/tpe/oo/metropolis/Steuerpflichtig.java
@@ -1,5 +1,5 @@
 package tpe.oo.metropolis;
 
 public interface Steuerpflichtig {
-double steuer();
+int steuer();
 }
