+++ 04_objektorientierung/13_solid/src/tpe/oo/metropolis/Superheld.java
@@ -2,7 +2,7 @@ package tpe.oo.metropolis;
 
 public class Superheld extends Mutant {
 
-    protected Superheld(String name, double einkommen, String mutation) {
+    protected Superheld(String name, int einkommen, String mutation) {
         super(name, einkommen, mutation);
 
     }
